myproject6/requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
python-dotenv
gunicorn